pyproject.toml
[project]
name = "disclaude-gate"
version = "0.1.0"
description = "Remote approval for Claude Code via Discord — approve, deny, or reply from your phone"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.3,<3",
    "aiohttp>=3.9,<4",
]

[project.scripts]
disclaude-gate = "src.server:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"
